remove trailing whitespace from nats.conf in ConfigMap (#296)
Depending on your configuration, the configmap can be rendered with
lines having trailing whitespace. This causes k8s to display the
configmap as just a big string containing literal "\n" strings, rather
than actual linebreaks.

This is most notably a problem when using `kubectl edit`, as it is
somewhat more difficult to edit the config when it's all on one line.

https://github.com/kubernetes/kubernetes/issues/36222 has a whole bunch
of discussion on the issue, in case you wanted to read about a bunch of
people yelling past each other.

Regardless of whether it's a bug or not, seems best to just avoid it in
the first place.