Support different k8s clusters on nats helm chart
It's not possible to create a nats cluster which includes nats nodes
on different kubernetes clusters.
Using the helm chart, I cannot explicitly set the cluster routes expected
in the configMap since it generates it based on the current k8s cluster.
This commit makes it possible with a new value nats.extraRoutes and
nats.serverName.