FIX Add missing lastPerSubject deliverPolicy option for Consumer CRD
Support the 'lastPerSubject' option for DeliverPolicy when creating a consumer
as documented in NATS JetStream documentation. This option was missing from
the current v1beta2 CRD schema, causing validation errors when trying to
use this feature:

'spec.deliverPolicy: Unsupported value: "lastPerSubject": supported values:
"all", "last", "new", "byStartSequence", "byStartTime"'

This option is named differently than in the NATS documentation, where
it is called `DeliverLastPerSubject`, or in the `nats consumer info -j`
output, where it is `last_per_subject`. The camel-case naming is based
on the NACK implementation in `consumerSpecToOpts()` in the natsio/nack
repo file `controllers/jetstream/consumer.go` which processes
`deliverPolicy` values.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>